chore: split student and worksheet routes out of index.ts
Pure mechanical extraction, no behavior change. Continues the split
started in batch-a/batch-b.

- routes/students.ts - GET/POST /api/students, GET
  /api/students/:id/diagnostic-paper, PATCH /api/students/:id[/profile],
  POST /api/students/:id/diagnostic[/submit], POST /api/paper/generate.
  This is the largest single-domain group so far (Python subprocess
  pipeline + Puppeteer paper generation + Gemini fallback all included
  unchanged).
- routes/worksheets.ts - POST /api/worksheets/generate[-pdf]
  [-level-pdf][-level-batch], GET /api/worksheets/download-batch/:id.
  Includes the generateLevelWorksheetsViaLevelsBackend shared helper
  (previously a free function inside startServer()), moved alongside
  the two routes that call it.

Verified: tsc --noEmit clean; booted the real server against a scratch
MongoDB seed. Tested the actual pipelines, not just route-exists 200s:
GET /api/students + /api/students/:id/diagnostic-paper both 200 with
real data; POST /api/worksheets/generate against a real class ran the
full personalized-question-compilation pipeline end-to-end and
returned a genuine 20-student worksheet with correctly per-student-
tagged question IDs; GET /api/worksheets/download-batch/:id correctly
502s on a fake batch id (route reached, external service call made,
not a 404/500 crash).

Closes #132, closes #131
