+++ backend/src/index.ts
@@ -31,6 +31,8 @@ import { registerTeacherRoutes } from './routes/teachers';
 import { registerSchoolRoutes } from './routes/schools';
 import { registerInterventionRoutes } from './routes/interventions';
 import { registerBestPracticeRoutes } from './routes/bestPractices';
+import { registerStudentRoutes } from './routes/students';
+import { registerWorksheetRoutes } from './routes/worksheets';
 import { randomUUID } from 'crypto';
 import fs from 'fs';
 import bcrypt from 'bcrypt';
@@ -75,557 +77,7 @@ registerStatsRoutes(app);
   // Classes
   registerClassRoutes(app);
 
-  // Students
-  app.get('/api/students', async (req, res) => {
-      const user = getAuthUser(req);
-      if (!user) return res.status(401).json({ error: 'Unauthorized' });
-
-      // The students collection has 86400+ docs in Atlas; without a server-side
-      // limit a single query takes multi-seconds and the dashboard hangs. Push the
-      // limit/offset into mongo. Default 1000 unless caller opts in to full set.
-      const DEFAULT_LIMIT = 1000;
-      const requestedLimit = parseInt(String(req.query.limit ?? ''), 10);
-      const requestedOffset = parseInt(String(req.query.offset ?? ''), 10) || 0;
-      const wantAll = req.query.all === '1' || req.query.all === 'true';
-      const limit = wantAll ? 0 : (Number.isFinite(requestedLimit) && requestedLimit > 0 ? Math.min(requestedLimit, DEFAULT_LIMIT * 5) : DEFAULT_LIMIT);
-
-      // server-side role scoping
-      let schoolScope: string | undefined;
-      if (user.role === UserRole.TEACHER || user.role === UserRole.SCHOOL) {
-        schoolScope = user.schoolId;
-      }
-
-      const opts: { limit?: number; offset?: number; schoolId?: string } = {
-        offset: requestedOffset,
-      };
-      if (limit > 0) opts.limit = limit;
-      if (schoolScope) opts.schoolId = schoolScope;
-
-      const students = await dbStore.getStudents(opts);
-
-      // volunteer filter still applied in JS (assignedSchools list, not a single key)
-      const filtered = (user.role === UserRole.VOLUNTEER)
-        ? students.filter(s => user.assignedSchools?.includes(s.schoolId))
-        : students;
-
-      // Mask Aadhar for non-Superadmins (§13.2 R-6)
-      const masked = filtered.map(s => {
-        if (user.role !== UserRole.SUPERADMIN) {
-          return { ...s, aadharMasked: 'XXXX-XXXX-' + String(s.aadharMasked || '').slice(-4) };
-        }
-        return s;
-      });
-
-      // total count (for client-side pagination headers)
-      const total = await dbStore.countStudents(schoolScope ? { schoolId: schoolScope } : undefined);
-      res.set('X-Total-Count', String(total));
-      res.json(masked);
-    });
-
-  // Get or generate student's assigned 10-question FLN paper from MongoDB Atlas (Class 2: Levels 22 to 31)
-  app.get('/api/students/:id/diagnostic-paper', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user) return res.status(401).json({ error: 'Unauthorized' });
-
-const students = await dbStore.getStudents();
-
-    // Roles with a direct, day-to-day relationship to the child (and superadmin)
-    // see full contact/address PII; aggregate-scope admins and volunteers get it
-    // redacted — they don't need a guardian's phone number to view rollups.
-    const canSeeGuardianPII = (role: UserRole) =>
-      role === UserRole.SUPERADMIN || role === UserRole.SCHOOL || role === UserRole.TEACHER;
-
-    // Mask Aadhar for non-Superadmins (§13.2 R-6); redact guardian contact/address similarly.
-    const maskedStudents = students.map(s => {
-      const masked = user.role !== UserRole.SUPERADMIN
-        ? { ...s, aadharMasked: 'XXXX-XXXX-' + s.aadharMasked.slice(-4) }
-        : { ...s };
-      if (!canSeeGuardianPII(user.role)) {
-        delete masked.guardianContact;
-        delete masked.address;
-      }
-      return masked;
-    });
-
-    let scoped: typeof maskedStudents;
-    if (user.role === UserRole.SUPERADMIN) {
-      scoped = students;
-    } else if (user.role === UserRole.SCHOOL || user.role === UserRole.TEACHER) {
-      scoped = maskedStudents.filter(s => s.schoolId === user.schoolId);
-    } else if (user.role === UserRole.VOLUNTEER) {
-      scoped = maskedStudents.filter(s => user.assignedSchools?.includes(s.schoolId));
-    } else if (user.role === UserRole.ADMIN || user.role === UserRole.DISTRICT_ADMIN || user.role === UserRole.BLOCK_ADMIN) {
-      // Geo-scope by the admin's own state/district/block, joined via each student's school.
-      const schools = await dbStore.getSchools();
-      const schoolById = new Map(schools.map(sc => [sc.id, sc]));
-      scoped = maskedStudents.filter(s => {
-        const school = schoolById.get(s.schoolId);
-        if (!school) return false;
-        if (user.role === UserRole.ADMIN) return school.stateCode === user.stateCode;
-        if (user.role === UserRole.DISTRICT_ADMIN) return school.districtCode === user.districtCode;
-        return school.blockCode === user.blockCode; // BLOCK_ADMIN
-      });
-    } else {
-      scoped = maskedStudents;
-    }
-
-    // Pagination is opt-in via ?page & ?limit — omitting them returns the full
-    // scoped array exactly as before, so existing callers (aggregate/rollup
-    // panels that need the whole scope) are unaffected. Callers that just need
-    // a page to display (the Student List table) can request one directly
-    // instead of always paying for the full national fetch.
-    const pageParam = req.query.page as string | undefined;
-    const limitParam = req.query.limit as string | undefined;
-    if (pageParam || limitParam) {
-      const page = Math.max(1, parseInt(pageParam || '1', 10) || 1);
-      const limit = Math.max(1, Math.min(500, parseInt(limitParam || '50', 10) || 50));
-      const total = scoped.length;
-      const start = (page - 1) * limit;
-      res.set('X-Total-Count', String(total));
-      res.set('X-Page', String(page));
-      res.set('X-Pages', String(Math.max(1, Math.ceil(total / limit))));
-      return res.json(scoped.slice(start, start + limit));
-    }
-
-    res.json(scoped);
-  });
-
-  // Add Student
-  app.post('/api/students', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user) return res.status(401).json({ error: 'Unauthorized' });
-
-    const {
-      name, age, classGroup, section, schoolId, aadharNumber,
-      gender, dob, guardianName, guardianRelation, guardianContact, address,
-      bloodGroup, disabilityStatus, midDayMealBeneficiary, busRoute, siblingsInSchool,
-    } = req.body;
-    if (!name || !age || !classGroup || !section || !schoolId || !aadharNumber) {
-      return res.status(400).json({ error: 'Missing required student details.' });
-    }
-
-    // Enforce Aadhar formatting & masking (§13.2 R-6)
-    const rawAadhar = aadharNumber.replace(/[^0-9]/g, '');
-    if (rawAadhar.length < 4) {
-      return res.status(400).json({ error: 'Invalid identity document.' });
-    }
-
-    // Enforce uniqueness check on raw Aadhar number
-    const studentsListForDuplicateCheck = await dbStore.getStudents();
-    const isDuplicate = studentsListForDuplicateCheck.some(s => s.aadharMasked === rawAadhar);
-    if (isDuplicate) {
-      return res.status(400).json({ error: 'A student with this Aadhar / ID number is already registered.' });
-    }
-
-    const newStudent: Student = {
-      id: 'STD_' + Math.floor(10000 + Math.random() * 90000),
-      name,
-      age: parseInt(age),
-      classGroup,
-      section,
-      schoolId,
-      teacherId: user.role === UserRole.TEACHER ? user.id : undefined,
-      currentLevel: 1, // Start at level 1 before diagnostic
-      currentSubLevel: 0,
-      targetLevel: 2,
-      aadharMasked: rawAadhar, // Store raw unmasked Aadhar in DB so Superadmin sees it, others get masked dynamically
-      levelHistory: [],
-      gender: gender || undefined,
-      dob: dob || undefined,
-      guardianName: guardianName || undefined,
-      guardianRelation: guardianRelation || undefined,
-      guardianContact: guardianContact || undefined,
-      address: address || undefined,
-      bloodGroup: bloodGroup || undefined,
-      disabilityStatus: disabilityStatus || undefined,
-      midDayMealBeneficiary: midDayMealBeneficiary === undefined ? undefined : Boolean(midDayMealBeneficiary),
-      busRoute: busRoute || undefined,
-      siblingsInSchool: siblingsInSchool || undefined,
-    };
-
-    await dbStore.addStudent(newStudent);
-
-    await dbStore.addLog({
-      id: 'log_' + Date.now(),
-      timestamp: new Date().toISOString(),
-      schoolId: schoolId,
-      schoolName: 'GPS',
-      userId: user.id,
-      userEmail: user.email,
-      userRole: user.role,
-      activityType: 'verify',
-      status: 'Success',
-      details: `Onboarded and verified student: ${name}`
-    });
-
-    res.json(newStudent);
-  });
-
-  // Update Student (Bypass / manual override for demo ease)
-  app.patch('/api/students/:id', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user) return res.status(401).json({ error: 'Unauthorized' });
-
-    const { currentLevel, currentSubLevel, targetLevel, levelHistory } = req.body;
-    const students = await dbStore.getStudents();
-    const student = students.find(s => s.id === req.params.id);
-    if (!student) return res.status(404).json({ error: 'Student not found.' });
-    if (!canAccessStudent(user, student)) return res.status(403).json({ error: 'Forbidden.' });
-
-    await dbStore.updateStudent(student.id, {
-      currentLevel: Number(currentLevel),
-      currentSubLevel: currentSubLevel !== undefined ? Number(currentSubLevel) : student.currentSubLevel,
-      targetLevel: Number(targetLevel),
-      levelHistory: levelHistory || student.levelHistory
-    });
-
-    res.json({ success: true });
-  });
-
-  // Update Student Profile (guardian/medical/logistics fields) — only the
-  // student's own school/teacher, or higher admins, may edit; kept separate
-  // from the level-update PATCH above so neither contract has to change.
-  app.patch('/api/students/:id/profile', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user) return res.status(401).json({ error: 'Unauthorized' });
-
-    const students = await dbStore.getStudents();
-    const student = students.find(s => s.id === req.params.id);
-    if (!student) return res.status(404).json({ error: 'Student not found.' });
-    if (!canAccessStudent(user, student)) return res.status(403).json({ error: 'Forbidden.' });
-
-    const {
-      gender, dob, guardianName, guardianRelation, guardianContact, address,
-      bloodGroup, disabilityStatus, midDayMealBeneficiary, busRoute, siblingsInSchool, teacherNotes,
-    } = req.body;
-
-    const updates: Partial<Student> = {};
-    if (gender !== undefined) updates.gender = gender;
-    if (dob !== undefined) updates.dob = dob;
-    if (guardianName !== undefined) updates.guardianName = guardianName;
-    if (guardianRelation !== undefined) updates.guardianRelation = guardianRelation;
-    if (guardianContact !== undefined) updates.guardianContact = guardianContact;
-    if (address !== undefined) updates.address = address;
-    if (bloodGroup !== undefined) updates.bloodGroup = bloodGroup;
-    if (disabilityStatus !== undefined) updates.disabilityStatus = disabilityStatus;
-    if (midDayMealBeneficiary !== undefined) updates.midDayMealBeneficiary = Boolean(midDayMealBeneficiary);
-    if (busRoute !== undefined) updates.busRoute = busRoute;
-    if (siblingsInSchool !== undefined) updates.siblingsInSchool = siblingsInSchool;
-    if (teacherNotes !== undefined) updates.teacherNotes = teacherNotes;
-
-    await dbStore.updateStudent(student.id, updates);
-
-    res.json({ success: true });
-  });
-
-  // Run Onboarding AI Diagnostic Test
-  app.post('/api/students/:id/diagnostic', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user) return res.status(401).json({ error: 'Unauthorized' });
-
-    const students = await dbStore.getStudents();
-    const student = students.find(s => s.id === req.params.id);
-    if (!student) return res.status(404).json({ error: 'Student not found.' });
-    if (!canAccessStudent(user, student)) return res.status(403).json({ error: 'Forbidden.' });
-
-    // Parse class number from classGroup (e.g. "Class 2" -> 2)
-    const classMatch = student.classGroup.match(/\d+/);
-    const classNumber = classMatch ? parseInt(classMatch[0], 10) : 1;
-
-    let questions: Question[];
-    let pdfUrl = '';
-
-    try {
-      // Generate the official PDF worksheet paper via Puppeteer
-      const result = await generateDiagnosticPaper({
-        classNumber,
-        students: [{
-          name: student.name,
-          studentId: student.id,
-          qrData: {
-            age: student.age, classGroup: student.classGroup, section: student.section,
-            schoolId: student.schoolId, currentLevel: student.currentLevel,
-            currentSubLevel: student.currentSubLevel, targetLevel: student.targetLevel
-          }
-        }]
-      });
-      questions = result.questions;
-      pdfUrl = `/output/${result.fileName}`;
-    } catch (err: any) {
-      console.error("Puppeteer paper generation failed, using level generator mock:", err);
-      const startLevel = (classNumber - 1) * 12 + 1;
-      questions = [];
-      for (let lvl = startLevel; lvl < startLevel + 8; lvl++) {
-        const lvlQuestions = generateQuestionsForLevel(Math.min(lvl, 93), 0);
-        lvlQuestions.forEach(q => {
-          questions.push({
-            ...q,
-            question_id: `DIAG_${lvl}_${q.question_id}`,
-            source_level: Math.min(lvl, 93)
-          });
-        });
-      }
-      questions = questions.slice(0, 12);
-    }
-
-    res.json({
-      student,
-      diagnosticPaper: {
-        id: 'diag_' + student.id + '_' + Date.now(),
-        studentId: student.id,
-        studentName: student.name,
-        questions,
-        pdfUrl
-      }
-    });
-  });
-
-  // Generate multi-student PDF worksheet paper (Puppeteer pipeline)
-  app.post('/api/paper/generate', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user) return res.status(401).json({ error: 'Unauthorized' });
-
-    try {
-      const { class: classNumber, students } = req.body;
-      if (classNumber === undefined || classNumber === null) {
-        return res.status(400).json({ success: false, error: 'class is required.' });
-      }
-      if (!Array.isArray(students) || students.length === 0) {
-        return res.status(400).json({ success: false, error: 'students must be a non-empty array.' });
-      }
-
-      const result = await generateDiagnosticPaper({
-        classNumber: Number(classNumber),
-        students: students.map((s: any) => ({ ...s, studentId: s.studentId || s.id || s.rollNo }))
-      });
-
-      const pdfUrl = `/output/${result.fileName}`;
-      res.json({
-        success: true,
-        pdfUrl,
-        totalSets: result.totalSets,
-        studentOrder: result.studentOrder
-      });
-    } catch (err: any) {
-      console.error('Failed to generate class paper sets:', err);
-      res.status(500).json({ success: false, error: err.message });
-    }
-  });
-
-  // Submit and evaluate Diagnostic responses
-  app.post('/api/students/:id/diagnostic/submit', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user) return res.status(401).json({ error: 'Unauthorized' });
-
-    const { questions, answers } = req.body;
-    const students = await dbStore.getStudents();
-    const student = students.find(s => s.id === req.params.id);
-    if (!student) return res.status(404).json({ error: 'Student not found.' });
-    if (!canAccessStudent(user, student)) return res.status(403).json({ error: 'Forbidden.' });
-
-    // Parse class number from classGroup (e.g. "Class 2" -> 2)
-    const classMatch = student.classGroup.match(/\d+/);
-    const classNumber = classMatch ? parseInt(classMatch[0], 10) : 1;
-
-    const dateStr = new Date().toISOString().split('T')[0];
-
-    // Idempotency: if this student's diagnostic was already submitted and
-    // evaluated today (e.g. a client retry after a timeout), return that
-    // existing report instead of re-running the pipeline and re-appending to
-    // level history. A genuinely new diagnostic on a later date still runs
-    // normally (legitimate re-assessment, not a duplicate retry).
-    const existingReports = await dbStore.getEvaluationReports();
-    const existingReport = existingReports.find(r =>
-      r.worksheetId === 'diagnostic' && r.studentId === student.id && r.timestamp.startsWith(dateStr)
-    );
-    if (existingReport) {
-      return res.json({
-        student,
-        evaluation: { score: existingReport.score, recommendedLevel: existingReport.recommendedLevel, narrative: existingReport.narrative },
-        report: existingReport,
-        alreadySubmitted: true
-      });
-    }
-
-    // Connect to Python Evaluation Metrics Pipeline
-    const pipelineDir = AI_SERVICES_DIR;
-    const responseDir = path.join(pipelineDir, 'student_responses', `class_${classNumber}`, 'phrase_1');
-    fs.mkdirSync(responseDir, { recursive: true });
-
-    // Map answers sequentially (diag_q_X_Y to Q1, Q2, Q3...)
-    const pipelineAnswers: { [qId: string]: { answer: string, confidence: number } } = {};
-    questions.forEach((q, idx) => {
-      const qNum = idx + 1;
-      const pipelineQId = `Q${qNum}`;
-      const submitted = (answers[q.question_id] || '').trim();
-      pipelineAnswers[pipelineQId] = {
-        answer: String(submitted),
-        confidence: 0.95
-      };
-    });
-
-    const studentResponse = {
-      student_id: student.id,
-      student_name: student.name,
-      enrolled_class: classNumber,
-      test_date: dateStr,
-      phrase: 'phrase_1',
-      exam_id: `C${classNumber}_WORKSHEET_PHRASE_1`,
-      answers: pipelineAnswers
-    };
-
-    const responsePath = path.join(responseDir, `${student.id}.json`);
-    fs.writeFileSync(responsePath, JSON.stringify(studentResponse, null, 2));
-
-    let score = 0;
-    let recommendedLevel = 1;
-    let narrative = '';
-
-    try {
-      const { execFileSync } = await import('child_process');
-      console.log(`Running evaluation pipeline for student ${student.id}...`);
-
-      // Run the comparison, evaluation, and report card generation pipeline.
-      // execFile (no shell) with array args means classNumber/student.id are passed
-      // literally and can never be interpreted as shell syntax.
-      execFileSync(PYTHON_BIN, ['run_pipeline.py', String(classNumber), 'phrase_1', student.id], {
-        cwd: pipelineDir,
-        env: { ...process.env, PYTHONIOENCODING: 'utf-8' }
-      });
-
-      // If failed, run the personalized exam pipeline too
-      try {
-        execFileSync(PYTHON_BIN, ['personalized_evaluation_pipeline.py', student.id, String(classNumber), 'phrase_1'], {
-          cwd: pipelineDir,
-          env: { ...process.env, PYTHONIOENCODING: 'utf-8' }
-        });
-      } catch (pexErr) {
-        console.warn('Personalized exam generation skipped or failed:', pexErr);
-      }
-
-      // Read evaluation result JSON and report text
-      const evalReportPath = path.join(pipelineDir, 'evaluation_reports', `class_${classNumber}`, 'phrase_1', 'evaluation', `${student.id}_evaluation_${dateStr}.json`);
-      const reportTxtPath = path.join(pipelineDir, 'evaluation_reports', `class_${classNumber}`, 'phrase_1', 'reports', `${student.id}_report_${dateStr}.txt`);
-
-      if (fs.existsSync(evalReportPath)) {
-        const evalData = JSON.parse(fs.readFileSync(evalReportPath, 'utf-8'));
-        score = evalData.total_questions - (evalData.wrong_count || 0);
-
-        const levelStr = String(evalData.demonstrated_level || '1');
-        const lvlMatch = levelStr.match(/\d+/);
-        if (lvlMatch) {
-          const matchedNum = parseInt(lvlMatch[0], 10);
-          if (levelStr.toLowerCase().includes('class')) {
-            recommendedLevel = (matchedNum - 1) * 10 + 1;
-          } else {
-            recommendedLevel = matchedNum;
-          }
-        } else {
-          recommendedLevel = 1;
-        }
-      }
-
-      if (fs.existsSync(reportTxtPath)) {
-        narrative = fs.readFileSync(reportTxtPath, 'utf-8');
-      }
-    } catch (pipelineErr) {
-      console.error('Python evaluation pipeline failed, falling back to Gemini AI:', pipelineErr);
-      // Fallback to Gemini AI if Python pipeline fails
-      const evaluation = await evaluateAIDiagnostic(student.name, questions, answers);
-      score = evaluation.score;
-      recommendedLevel = evaluation.recommendedLevel;
-      narrative = evaluation.narrative;
-    }
-
-    // Determine the subLevel based on weakest-level mapping questions
-    let subLevel = 0; // default Mastery
-    const levelQuestions = questions.filter(q => q.source_level === recommendedLevel);
-    if (levelQuestions.length > 0) {
-      let failedCount = 0;
-      levelQuestions.forEach(q => {
-        const submitted = (answers[q.question_id] || '').trim().toLowerCase();
-        const correct = q.answer.trim().toLowerCase();
-        if (submitted !== correct) {
-          failedCount++;
-        }
-      });
-
-      if (failedCount === levelQuestions.length) {
-        subLevel = 2; // Remedial (failed all)
-      } else if (failedCount > 0) {
-        subLevel = 1; // Easier (failed some)
-      } else {
-        subLevel = 0; // Mastery
-      }
-    }
-
-    // Update Student placing levels
-    const levelHistory = [...student.levelHistory, {
-      level: recommendedLevel,
-      subLevel,
-      date: new Date().toISOString().split('T')[0],
-      reason: CYCLE_NAMES[0] // 'Baseline'
-    }];
-
-    await dbStore.updateStudent(student.id, {
-      currentLevel: recommendedLevel,
-      currentSubLevel: subLevel,
-      targetLevel: Math.min(93, recommendedLevel + 1),
-      levelHistory
-    });
-
-    // Create a special Evaluation Report with dynamic mock concept mastery
-    const conceptMastery: { [topic: string]: "Strong" | "Needs Practice" | "Satisfactory" } = {
-      'Number Sense': recommendedLevel >= 15 ? 'Strong' : 'Needs Practice',
-      'Shapes': recommendedLevel >= 25 ? 'Strong' : 'Needs Practice',
-      'Fractions': recommendedLevel >= 35 ? 'Strong' : 'Needs Practice',
-      'Operations': recommendedLevel >= 12 ? 'Strong' : 'Needs Practice'
-    };
-
-    try {
-      const evalReportPath = path.join(pipelineDir, 'evaluation_reports', `class_${classNumber}`, 'phrase_1', 'evaluation', `${student.id}_evaluation_${dateStr}.json`);
-      if (fs.existsSync(evalReportPath)) {
-        const evalData = JSON.parse(fs.readFileSync(evalReportPath, 'utf-8'));
-        if (evalData.topics_to_focus && Array.isArray(evalData.topics_to_focus)) {
-          evalData.topics_to_focus.forEach((t: string) => {
-            conceptMastery[t] = 'Needs Practice';
-          });
-        }
-      }
-    } catch (e) {
-      console.warn('Failed to parse dynamic concept mastery:', e);
-    }
-
-    const report: EvaluationReport = {
-      id: 'rep_diag_' + Date.now(),
-      studentId: student.id,
-      worksheetId: 'diagnostic',
-      score,
-      totalQuestions: questions.length,
-      conceptMastery,
-      narrative,
-      recommendedLevel,
-      recommendedSubLevel: subLevel,
-      timestamp: new Date().toISOString()
-    };
-
-    await dbStore.addEvaluationReport(report);
-
-    await dbStore.addLog({
-      id: 'log_' + Date.now(),
-      timestamp: new Date().toISOString(),
-      schoolId: student.schoolId,
-      schoolName: 'GPS',
-      userId: user.id,
-      userEmail: user.email,
-      userRole: user.role,
-      activityType: 'scan',
-      status: 'Success',
-      details: `Submitted and scored diagnostic for ${student.name}. Placed at Level ${recommendedLevel}`
-    });
-
-    res.json({ student, evaluation: { score, recommendedLevel, narrative }, report });
-  });
+  registerStudentRoutes(app);
 
   // ICR Blue-Pen Filter Stage (standalone — runs only the cv2 blue-pen
   // isolation, no OCR). Returns the filtered image as a data URL so the
@@ -1387,440 +839,7 @@ const students = await dbStore.getStudents();
       return res.status(500).json({ error: 'Cloud OCR failed: ' + (e && e.message ? e.message : String(e)) });
     }
   });
-
-// Generate Personalized Class Worksheets
-  app.post('/api/worksheets/generate', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user) return res.status(401).json({ error: 'Unauthorized' });
-
-    const { classId, cycle } = req.body;
-    if (!classId || !cycle) {
-      return res.status(400).json({ error: 'Class ID and assessment cycle are required.' });
-    }
-
-    const classes = await dbStore.getClasses();
-    const classObj = classes.find(c => c.id === classId);
-    if (!classObj) return res.status(404).json({ error: 'Class not found.' });
-
-    // Check if school is low or high strength
-    const schools = await dbStore.getSchools();
-    const school = schools.find(s => s.id === classObj.schoolId);
-    if (!school) return res.status(404).json({ error: 'School not found.' });
-
-    // Check if Teacher is banned due to Delayed Attempts (§6.5)
-    if (user.role === UserRole.TEACHER && user.isBanned) {
-      return res.status(403).json({ error: 'Generation Denied: Teacher account is suspended/banned due to 3 Delayed Attempts within the academic year.' });
-    }
-
-    // Check if School is locked out entirely (§6.5)
-    if (school.isAccessLocked) {
-      if (user.role === UserRole.TEACHER || user.role === UserRole.SCHOOL) {
-        return res.status(403).json({ error: 'School Access Suspended: All teachers have defaulted. Management is reassigned to Block Admin / Volunteer.' });
-      }
-    }
-
-    // Check for Generation Lock (§13.2 R-11)
-    const existingWorksheets = await dbStore.getWorksheets();
-    const conflicting = existingWorksheets.find(w => w.classId === classId && w.cycle === cycle);
-
-    if (conflicting && conflicting.locks.locked) {
-      // Enforce pairwise lockouts
-      if (school.strength === 'high') {
-        // Teacher ↔ School pair
-        if (conflicting.locks.lockedByRole !== user.role) {
-          return res.status(423).json({
-            error: `Lock Active: Generation has already been triggered by ${conflicting.locks.lockedByRole} (${conflicting.locks.lockedByEmail}). Parallel generation is locked.`,
-            lockDetails: conflicting.locks
-          });
-        }
-      } else {
-        // Volunteer ↔ Block Admin pair
-        if (conflicting.locks.lockedByRole !== user.role) {
-          return res.status(423).json({
-            error: `Lock Active: Generation has already been triggered by ${conflicting.locks.lockedByRole} (${conflicting.locks.lockedByEmail}). Parallel generation is locked.`,
-            lockDetails: conflicting.locks
-          });
-        }
-      }
-    }
-
-    // Generate personalized questions for every student in the class
-    const students = await dbStore.getStudents();
-    const classStudents = students.filter(s => s.classGroup === classObj.className && s.section === classObj.section && s.schoolId === classObj.schoolId);
-
-    if (classStudents.length === 0) {
-      return res.status(400).json({ error: 'No students found in this class roster.' });
-    }
-
-    // Compile distinct personalized questions per student based on level and sub-level
-    const compiledQuestions: Question[] = [];
-
-    for (const student of classStudents) {
-      const subLvl = student.currentSubLevel || 0;
-      const qs = generateQuestionsForLevel(student.currentLevel, subLvl);
-      // Map question IDs to be student-specific to prevent duplicate collisions
-      qs.forEach(q => {
-        compiledQuestions.push({
-          ...q,
-          question_id: `${student.id}_${q.question_id}`,
-          question: `[For ${student.name} - L${student.currentLevel}.${subLvl}] ${q.question}`
-        });
-      });
-    }
-
-    // Setup strict Timing Windows (§1.4 Sequential timings)
-    const now = new Date();
-    const todayStr = now.toISOString().split('T')[0];
-
-    // Check if other worksheets exist for the same school on the same day to make print windows sequential & non-overlapping
-    const sameDayWorksheets = existingWorksheets.filter(w => w.schoolId === classObj.schoolId && w.date === todayStr);
-
-    let printStart = new Date(now.getTime());
-    if (sameDayWorksheets.length > 0) {
-      // Find the latest printWindowEnd
-      const latestEnd = new Date(Math.max(...sameDayWorksheets.map(w => new Date(w.timing.printWindowEnd).getTime())));
-      if (latestEnd.getTime() > now.getTime()) {
-        printStart = latestEnd;
-      }
-    }
-
-    const printEnd = new Date(printStart.getTime() + 60 * 60 * 1000); // 1 hour print window
-    const examStart = new Date(printEnd.getTime());
-    const examEnd = new Date(examStart.getTime() + 45 * 60 * 1000); // 45 mins exam
-    const submissionEnd = new Date(examEnd.getTime() + 60 * 60 * 1000); // 1 hour upload
-
-    const newWorksheet: Worksheet = {
-      id: 'WS_' + Math.floor(1000 + Math.random() * 9000),
-      classId,
-      className: classObj.className,
-      section: classObj.section,
-      schoolId: classObj.schoolId,
-      generatedByRole: user.role,
-      generatedByEmail: user.email,
-      cycle,
-      date: todayStr,
-      questions: compiledQuestions,
-      locks: {
-        locked: true,
-        lockedByRole: user.role,
-        lockedByEmail: user.email,
-        timestamp: now.toISOString()
-      },
-      timing: {
-        examDate: todayStr,
-        printWindowStart: printStart.toISOString(),
-        printWindowEnd: printEnd.toISOString(),
-        examWindowStart: examStart.toISOString(),
-        examWindowEnd: examEnd.toISOString(),
-        submissionWindowEnd: submissionEnd.toISOString()
-      },
-      delayLogs: {
-        delayedAttemptsCount: 0,
-        submittingTeachers: []
-      }
-    };
-
-    await dbStore.addWorksheet(newWorksheet);
-
-    await dbStore.addLog({
-      id: 'log_' + Date.now(),
-      timestamp: now.toISOString(),
-      schoolId: classObj.schoolId,
-      schoolName: school.name,
-      userId: user.id,
-      userEmail: user.email,
-      userRole: user.role,
-      activityType: 'download',
-      status: 'Success',
-      details: `Generated personalized worksheets for ${classObj.className} ${classObj.section}. Locked pairwise role.`
-    });
-
-    res.json(newWorksheet);
-  });
-
-  // Generate printable PDF for an existing worksheet (connects 93 FLN levels with diagnostic pipeline)
-  app.post('/api/worksheets/generate-pdf', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user) return res.status(401).json({ error: 'Unauthorized' });
-
-    const { worksheetId } = req.body;
-    if (!worksheetId) {
-      return res.status(400).json({ error: 'worksheetId is required.' });
-    }
-
-    const worksheets = await dbStore.getWorksheets();
-    const ws = worksheets.find(w => w.id === worksheetId);
-    if (!ws) return res.status(404).json({ error: 'Worksheet not found.' });
-
-    const students = await dbStore.getStudents();
-    const classStudents = students.filter(
-      s => s.classGroup === ws.className && s.section === ws.section && s.schoolId === ws.schoolId
-    );
-
-    const studentsWithQuestions = classStudents.map(s => {
-      const studentQuestions = ws.questions.filter(q => q.question_id.startsWith(s.id + '_'));
-      return {
-        studentId: s.id,
-        name: s.name,
-        currentLevel: s.currentLevel,
-        currentSubLevel: s.currentSubLevel || 0,
-        questions: studentQuestions
-      };
-    }).filter(s => s.questions.length > 0);
-
-    if (studentsWithQuestions.length === 0) {
-      return res.status(400).json({ error: 'No student questions found in this worksheet.' });
-    }
-
-    try {
-      const { renderWorksheetPdf } = await import('./paperGenerator');
-      const result = await renderWorksheetPdf({
-        worksheetId,
-        className: ws.className,
-        section: ws.section,
-        cycle: ws.cycle,
-        studentsWithQuestions
-      });
-      res.json({ success: true, pdfUrl: result.pdfUrl });
-    } catch (err: any) {
-      console.error('Worksheet PDF generation failed:', err);
-      res.status(500).json({ success: false, error: err.message });
-    }
-  });
-
-  /**
-   * Shared pipeline: build a roster -> Levels_backend /api/generate-batch ->
-   * poll /api/batch-status -> /api/download-batch (zip) -> unpack
-   * worksheet.pdf + answer_key.json + coords.json -> save PDF into this
-   * backend's own /output (served statically) and persist a LevelWorksheet
-   * record (with the real answer key + OMR coords) per rendered file.
-   *
-   * Students are matched back to rendered files via the roster's
-   * `rollNumber` field, which we deliberately set to the student's stable
-   * internal id (this codebase has no separate roll-number field on
-   * Student) — Levels_backend echoes the original rollNumber back in its
-   * manifest.json per file, so the mapping is exact regardless of how it
-   * sanitizes folder names.
-   */
-  async function generateLevelWorksheetsViaLevelsBackend(
-    students: Student[],
-    _opts: { includeBatchId?: boolean } = {}
-  ): Promise<Array<{
-    studentId: string;
-    studentName: string;
-    batchId: string;
-    sublevelId: string;
-    setNum: number;
-    pdfUrl: string;
-  }>> {
-    const roster: levelsBackendClient.RosterEntry[] = students.map(s => ({
-      studentName: s.name,
-      rollNumber: s.id,
-      levelId: s.currentLevel,
-      sublevelId: s.currentSubLevel != null ? `${s.currentLevel}.${s.currentSubLevel}` : 'all',
-      setsPerSub: 1,
-      studentData: {
-        age: s.age, classGroup: s.classGroup, section: s.section, schoolId: s.schoolId,
-        currentLevel: s.currentLevel, currentSubLevel: s.currentSubLevel,
-        targetLevel: s.targetLevel
-      }
-    }));
-
-    const batchResult = await levelsBackendClient.generateBatch(roster);
-    await levelsBackendClient.waitForBatch(batchResult.batchId);
-    const zipBuffer = await levelsBackendClient.downloadBatchZip(batchResult.batchId);
-    const { manifest, files } = await levelsBackendClient.extractBatchZip(zipBuffer);
-
-    // groupKey ("<studentFolder>/<sublevelId>_set<n>") -> original rollNumber (== studentId)
-    const rollNumberByGroupKey = new Map<string, string>();
-    if (manifest && Array.isArray(manifest.students)) {
-      for (const ms of manifest.students) {
-        if (!Array.isArray(ms.files)) continue;
-        for (const f of ms.files) {
-          rollNumberByGroupKey.set(f.folder, ms.rollNumber);
-        }
-      }
-    }
-
-    const studentsById = new Map(students.map(s => [s.id, s]));
-    const localOutputDir = path.join(ROOT_DIR, 'output');
-    if (!fs.existsSync(localOutputDir)) fs.mkdirSync(localOutputDir, { recursive: true });
-
-    const out: Array<{ studentId: string; studentName: string; batchId: string; sublevelId: string; setNum: number; pdfUrl: string }> = [];
-
-    for (const file of files) {
-      const groupKey = `${file.studentFolder}/${file.sublevelId}_set${file.setNum}`;
-      const studentId = rollNumberByGroupKey.get(groupKey);
-      const student = studentId ? studentsById.get(studentId) : undefined;
-      if (!student) {
-        console.warn(`[levels-backend] Could not map rendered file back to a student: ${groupKey}`);
-        continue;
-      }
-
-      const fileName = `level_${student.currentLevel}_${file.sublevelId}_set${file.setNum}_${student.id}_${randomUUID()}.pdf`;
-      const filePath = path.join(localOutputDir, fileName);
-      fs.writeFileSync(filePath, file.pdfBuffer);
-      const pdfUrl = `/output/${fileName}`;
-
-      // Write corresponding JSONs alongside the PDF for single/batch files
-      const baseName = fileName.replace(/\.pdf$/, '');
-      if (file.answerKey) {
-        fs.writeFileSync(path.join(localOutputDir, `${baseName}_answer_key.json`), JSON.stringify(file.answerKey, null, 2));
-      }
-      if (file.coords) {
-        fs.writeFileSync(path.join(localOutputDir, `${baseName}_coords.json`), JSON.stringify(file.coords, null, 2));
-      }
-      if (file.questionPaper) {
-        fs.writeFileSync(path.join(localOutputDir, `${baseName}_question_paper.json`), JSON.stringify(file.questionPaper, null, 2));
-      }
-
-      const record: LevelWorksheet = {
-        id: 'LW_' + randomUUID(),
-        batchId: batchResult.batchId,
-        studentId: student.id,
-        studentName: student.name,
-        rollNumber: student.id,
-        levelId: student.currentLevel,
-        sublevelId: file.sublevelId,
-        setNum: file.setNum,
-        pdfUrl,
-        answerKey: file.answerKey,
-        coords: file.coords,
-        generatedAt: new Date().toISOString()
-      };
-      await dbStore.addLevelWorksheet(record);
-
-      out.push({
-        studentId: student.id,
-        studentName: student.name,
-        batchId: batchResult.batchId,
-        sublevelId: file.sublevelId,
-        setNum: file.setNum,
-        pdfUrl
-      });
-    }
-
-    return out;
-  }
-
-  // Generate Personalized Level-Wise Worksheet PDF for a single student.
-  // Pipeline: build a 1-entry roster -> Levels_backend /api/generate-batch
-  // -> poll /api/batch-status -> fetch /api/download-batch (zip) -> extract
-  // worksheet.pdf + answer_key.json + coords.json -> persist here.
-  app.post('/api/worksheets/generate-level-pdf', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user) return res.status(401).json({ error: 'Unauthorized' });
-
-    const { studentId } = req.body;
-    if (!studentId) {
-      return res.status(400).json({ error: 'studentId is required.' });
-    }
-
-    try {
-      const students = await dbStore.getStudents();
-      const student = students.find(s => s.id === studentId);
-      if (!student) return res.status(404).json({ error: 'Student not found.' });
-
-      if (student.currentLevel == null) {
-        return res.status(400).json({ error: 'Student has not completed their diagnostic test.' });
-      }
-
-      try {
-        const generated = await generateLevelWorksheetsViaLevelsBackend([student]);
-        if (generated.length === 0) {
-          throw new Error('Levels_backend returned no files for this student.');
-        }
-        res.json({ success: true, pdfUrl: generated[0].pdfUrl });
-      } catch (levelsBackendErr: any) {
-        // Deterministic fallback: the old in-process Puppeteer generator,
-        // so the button keeps working if Levels_backend is unreachable.
-        console.error('Levels_backend generation failed, falling back to local generator:', levelsBackendErr.message);
-        const { generateLevelWorksheet } = await import('./paperGenerator');
-        const result = await generateLevelWorksheet({
-          studentId: student.id,
-          studentName: student.name,
-          levelId: student.currentLevel,
-          subIdx: student.currentSubLevel || 0
-        });
-        res.json({ success: true, pdfUrl: result.pdfUrl, fallback: true });
-      }
-    } catch (err: any) {
-      console.error('Level worksheet generation failed:', err);
-      res.status(500).json({ success: false, error: err.message });
-    }
-  });
-
-  // Generate Personalized Level-Wise Worksheets for a whole roster of
-  // students in ONE batch call to Levels_backend (the "Generate Batch"
-  // button in the teacher dashboard's Level-Wise Paper Generator panel).
-  app.post('/api/worksheets/generate-level-batch', async (req, res) => {
-    const user = getAuthUser(req);
-    if (!user) return res.status(401).json({ error: 'Unauthorized' });
-
-    const { studentIds } = req.body;
-    if (!Array.isArray(studentIds) || studentIds.length === 0) {
-      return res.status(400).json({ error: 'studentIds must be a non-empty array.' });
-    }
-
-    try {
-      const students = await dbStore.getStudents();
-      const targets: Student[] = [];
-      const skipped: Array<{ studentId: string; reason: string }> = [];
-
-      for (const id of studentIds) {
-        const student = students.find(s => s.id === id);
-        if (!student) {
-          skipped.push({ studentId: id, reason: 'Student not found.' });
-          continue;
-        }
-        if (student.currentLevel == null) {
-          skipped.push({ studentId: id, reason: 'Student has not completed their diagnostic test.' });
-          continue;
-        }
-        targets.push(student);
-      }
-
-      if (targets.length === 0) {
-        return res.status(400).json({ error: 'No eligible (placed) students in this request.', skipped });
-      }
-
-      const generated = await generateLevelWorksheetsViaLevelsBackend(targets, { includeBatchId: true });
-
-      const results = generated.map(g => ({
-        studentId: g.studentId,
-        studentName: g.studentName,
-        sublevelId: g.sublevelId,
-        setNum: g.setNum,
-        pdfUrl: g.pdfUrl
-      }));
-
-      res.json({
-        success: true,
-        batchId: generated[0]?.batchId || null,
-        studentsProcessed: targets.length,
-        totalFiles: generated.length,
-        results,
-        skipped
-      });
-    } catch (err: any) {
-      console.error('Level-wise batch generation failed:', err);
-      res.status(500).json({ success: false, error: err.message });
-    }
-  });
-
-  // Streams the raw batch ZIP straight from Levels_backend, for the
-  // "Download Batch ZIP" button. No transformation — pass-through.
-  app.get('/api/worksheets/download-batch/:batchId', async (req, res) => {
-    try {
-      const zipBuffer = await levelsBackendClient.downloadBatchZip(req.params.batchId);
-      res.setHeader('Content-Type', 'application/zip');
-      res.setHeader('Content-Disposition', `attachment; filename="batch_${req.params.batchId}.zip"`);
-      res.send(zipBuffer);
-    } catch (err: any) {
-      console.error('Batch ZIP download failed:', err);
-      res.status(502).json({ error: err.message });
-    }
-  });
+  registerWorksheetRoutes(app);
 
   // Submit Completed Worksheet (ICR Structured Ingestion) & Scoring Engine
   app.post('/api/evaluation/submit', async (req, res) => {
